requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
mcp
clickhouse-connect
mcp[cli]
ddgs
SpeechRecognition
pyaudio
pvporcupine
pvrecorder
simpleaudio
numpy
packaging
torch
torchaudio<2.9
rumps
pystray
faster_whisper
supermemory
edge-tts
grpcio
grpcio-tools